Rename to aarch64-unknown-switch-libnx
