+++ library/std/src/os/libnx/mod.rs
@@ -1,4 +1,4 @@
-//! devkitA64-specific definitions
+//! libnx-specific definitions
 
 #![stable(feature = "raw_ext", since = "1.1.0")]
