+++ library/std/src/sys/unix/process/mod.rs
@@ -3,13 +3,13 @@ pub use self::process_inner::{ExitStatus, Process};
 pub use crate::ffi::OsString as EnvKey;
 
 mod process_common;
-#[cfg(not(any(target_os = "fuchsia", target_env = "devkita64")))]
+#[cfg(not(any(target_os = "fuchsia", target_env = "libnx")))]
 #[path = "process_unix.rs"]
 mod process_inner;
 #[cfg(target_os = "fuchsia")]
 #[path = "process_fuchsia.rs"]
 mod process_inner;
-#[cfg(target_env = "devkita64")]
+#[cfg(target_env = "libnx")]
 #[path = "process_dummy.rs"]
 mod process_inner;
 #[cfg(target_os = "fuchsia")]
