+++ library/std/src/sys/unix/process/process_dummy.rs
@@ -1,4 +1,4 @@
-//! devkitA64 has very limited support for multiple processes, so a dummy implementation is used
+//! libnx has very limited support for multiple processes, so a dummy implementation is used
 //! there (similar to WebAssembly).
 
 use crate::fmt;
