+++ src/tools/tidy/src/extdeps.rs
@@ -6,7 +6,7 @@ use std::path::Path;
 /// List of allowed sources for packages.
 const ALLOWED_SOURCES: &[&str] = &[
     "\"registry+https://github.com/rust-lang/crates.io-index\"",
-    "\"git+https://github.com/leo60228/libc.git?branch=bindgen-horizon#c64fc22e37bca14e942e0f7e8c365fc7b2d4c963\"",
+    "\"git+https://github.com/leo60228/libc.git?branch=bindgen-horizon#9db9538ce95d61b75964b3223cf03cafe1023327\"",
 ];
 
 /// Checks for external package sources. `root` is the path to the directory that contains the
